requirements.txt
pandas
praw
tqdm
pyyaml # only required if you will use the yaml file